pyproject.toml
[project]
name = "stvg-helper"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic",
    "av",
    "numpy",
    "onnxruntime",
    "Pillow",
    "python-telegram-bot==21.10",
    "requests",
]

[dependency-groups]
dev = [
    "boto3",
    "black",
    "isort",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "types-requests",
]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["bot"]
asyncio_mode = "auto"
